requirements.txt
# =========================
# Core Web Framework
# =========================
Flask==2.2.5
flask-cors==6.0.2
gunicorn==23.0.0
Werkzeug==2.2.3
Jinja2==3.1.3
itsdangerous==2.1.2
dlib-19.22.99-cp39-cp39-win_amd64.whl
# =========================
# AI / ML Stack
# =========================
tensorflow==2.15.0
numpy==1.23.5
scipy==1.10.1
pandas==2.3.3
scikit-learn==1.2.2
h5py==3.8.0
protobuf==4.25.8

# =========================
# Face Recognition
# =========================
deepface==0.0.79
face-recognition==1.3.0
face-recognition-models==0.3.0
mtcnn==0.1.1
retina-face==0.0.17
opencv-python==4.8.1.78
pillow==10.2.0

# =========================
# PyTorch (if required)
# =========================
torch==1.13.1
torchvision==0.14.1

# =========================
# Google Gemini / AI
# =========================
google-generativeai==0.5.1
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.62.3
proto-plus==1.27.1

# =========================
# Utilities
# =========================
requests==2.31.0
urllib3==1.26.18
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
PyYAML==6.0.1
tqdm==4.67.3
typing-extensions==4.15.0

# =========================
# Security
# =========================
cryptography==45.0.7
rsa==4.9.1
pyasn1==0.6.2
pyasn1-modules==0.4.2

# =========================
# File / Report Generation
# =========================
reportlab==4.4.10

# =========================
# Audio / Speech (Optional)
# =========================
SpeechRecognition==3.10.0
pydub==0.25.1
sounddevice==0.4.6
setuptools==68.0.0